+++ tests/integration/test_attachments.py
@@ -2,6 +2,7 @@ import pytest
 import uuid
 import json
 
+from unittest import mock
 from requests.exceptions import HTTPError
 from sentry_sdk.envelope import Envelope, Item, PayloadRef
 from objectstore_client import Client, Usecase
@@ -153,10 +154,10 @@ def test_attachments_with_objectstore(
     )
     assert objectstore_session.get(objectstore_key).payload.read() == chunked_contents
 
-    assert attachment["attachment"].pop("id")
     assert attachment == {
         "type": "attachment",
         "attachment": {
+            "id": mock.ANY,
             "name": "foo.txt",
             "rate_limited": False,
             "attachment_type": "event.attachment",
@@ -170,11 +171,10 @@ def test_attachments_with_objectstore(
 
     # An empty attachment
     attachment = attachments_consumer.get_individual_attachment()
-    assert attachment["attachment"].pop("id")
-
     assert attachment == {
         "type": "attachment",
         "attachment": {
+            "id": mock.ANY,
             "name": "foobar.txt",
             "rate_limited": False,
             "attachment_type": "event.attachment",